test(mint): move batch minting api tests to main api test file
